+++ HexToColor/Hex.swift
@@ -21,7 +21,7 @@ public func toColor(hex: String) -> UIColor {
         cString = cString.substringFromIndex(cString.startIndex.advancedBy(2))
     }
 
-    if (cString.characters.count == 3) {
+    if (cString.characters.count == 3 || cString.characters.count == 4) {
         var expandedString = ""
         for character in cString.characters {
             expandedString.append(character)
@@ -30,7 +30,7 @@ public func toColor(hex: String) -> UIColor {
         cString = expandedString
     }
     
-    if (cString.characters.count != 6) {
+    if (cString.characters.count != 6 && cString.characters.count != 8) {
         return UIColor.grayColor()
     }
     
@@ -39,11 +39,27 @@ public func toColor(hex: String) -> UIColor {
     if (!scanner.scanHexInt(&rgbValue) || !scanner.atEnd) {
         return UIColor.grayColor()
     }
+
+    let redValue: UInt32
+    let greenValue: UInt32
+    let blueValue: UInt32
+    let alphaValue: UInt32
+    if (cString.characters.count == 8) {
+        redValue = (rgbValue & 0xFF000000) >> 24
+        greenValue = (rgbValue & 0x00FF0000) >> 16
+        blueValue = (rgbValue & 0x0000FF00) >> 8
+        alphaValue = rgbValue & 0x000000FF
+    } else {
+        redValue = (rgbValue & 0xFF0000) >> 16
+        greenValue = (rgbValue & 0x00FF00) >> 8
+        blueValue = rgbValue & 0x0000FF
+        alphaValue = 0xFF
+    }
     
     return UIColor(
-        red: CGFloat((rgbValue & 0xFF0000) >> 16) / 255.0,
-        green: CGFloat((rgbValue & 0x00FF00) >> 8) / 255.0,
-        blue: CGFloat(rgbValue & 0x0000FF) / 255.0,
-        alpha: CGFloat(1.0)
+        red: CGFloat(redValue) / 255.0,
+        green: CGFloat(greenValue) / 255.0,
+        blue: CGFloat(blueValue) / 255.0,
+        alpha: CGFloat(alphaValue) / 255.0
     )
 }